change motion control to double; getLastFrameDuration
Former-commit-id: 13c95ade23ef26e71d5b585937b6c6e113bae4aa [formerly ef8e20a100c15863ce6d511a57d1c369403423d8] [formerly 0b91389749b574119b9c4174b26b225e20336b35 [formerly 515a7aa74296461df8509aca0cffb7a9add071e6]] [formerly 5eaf18543acd991733d03214a775473cc1604bc9 [formerly 006720a2951cf512c80b7679e8a008d1048cb1c6] [formerly cd2f95f5069bb9b7e528be3f2c1559bde2201df2 [formerly 68b2307a9bb2af5ac8b77923754c1cf3b82929b3]]] [formerly 72b1bf615eabd9f3b6e8bec99c2e5427697e4eef [formerly 579c008708aec9011377bc3378bd7fec7834476d] [formerly 45f1d27aaaa60e108b8e3aa090443797722fdd3c [formerly bb7e28b9b7eb19c6e85c724809f0fff082450fa3]] [formerly e3a3ff7fcb7de984399404ce0cc7eb1b78549b1c [formerly 57611d00d66aa129414ed91878a7415986d3dbed] [formerly 4451ce1501695e124fa1a7f0f5f01c3f79dcc24f [formerly fe125fd06a20f0c465e285aaf9cf074318f18da0]]]]
Former-commit-id: 9a7f0eb342cc80e53b056fd3705eeeed0842f59b [formerly 9b7b50683147ec9407cc27bf8aa1c4b45c6daa99] [formerly 266e8e9192bc5ff1893443e97db13c9a9f86688f [formerly 5c155b9bab91a2e5a041856005000db949dea559]] [formerly 2f3c5c2148e61f8f760378ed63202b5e04c7f9fe [formerly c993e72ee500ce8ef80856442e7adb6095e5ef3c] [formerly 7899d12e59d67d9b7f1e88d41dbbc4e3223b656a [formerly 79f3aba3c131e02bf114974664bf1ec78ca7623c]]]
Former-commit-id: 7a82867508c72fabc01ca2576002387d854850aa [formerly 47916440357c7f706fc40b329763222c60bcc4bd] [formerly c6ce906e03fec1fb6d6772953bb6ebbb414949db [formerly 2fed7652c76bc5d9caf910625886417494165647]]
Former-commit-id: cd68d8a7ff73b243c7c5c9e1e3f2d8a3a98f7c5e [formerly 131009a4025d713b06a8ac420327d3a04a7abbec]
Former-commit-id: ff693bfb46669799b4228dd25964f94fed2863da

+++ LiteEngine/IO/CameraController.h
@@ -142,27 +142,27 @@ namespace LiteEngine::IO {
 		}
 
 	public:
-		float accX = 0;
-		float accY = 0;
-		float accZ = 0;
+		double accX = 0;
+		double accY = 0;
+		double accZ = 0;
 
-		float xSensitivity = 0.003f;
-		float ySensitivity = 0.003f;
-		float zSensitivity = 0.003f;
+		double xSensitivity = 0.003f;
+		double ySensitivity = 0.003f;
+		double zSensitivity = 0.003f;
 
-		float Ymin = - float(0.9 * PI / 2);
-		float Ymax = + float(0.9 * PI / 2);
+		double Ymin = - double(0.9 * PI / 2);
+		double Ymax = + double(0.9 * PI / 2);
 
-		float zMin = -1;
-		float zMax = 1;
+		double zMin = -1;
+		double zMax = 1;
 
-		std::tuple<float, float, float> getXYZ() {
+		std::tuple<double, double, double> getXYZ() {
 			return { accX, accY, accZ };
 		}
 
 		void receiveEvent(
 			const std::vector<std::tuple<UINT, WPARAM, LPARAM>>& events, 
-			float lastFrameDuration
+			double lastFrameDuration
 		) {
 			for (auto& [msg, wparam, lparam] : events) {
 				switch (msg) {
